backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
fastapi==0.127.0
h11==0.16.0
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
requests==2.32.5
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uvicorn==0.40.0

# Database: PostgreSQL with PostGIS support
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
geoalchemy2>=0.15.0
python-dotenv==1.0.0

# Database migrations (Alembic)
alembic>=1.13.0

# Password hashing
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0

# Weather API integration
# OpenWeatherMap API is used via requests library (already included above)
